Add: Get an event from socket trigger canvas function
