+++ src/sockets/socketConnection.ts
@@ -1,42 +1,31 @@
-// import { io, Socket } from "socket.io-client";
-// import { getToken } from '@/utils/auth.utils';
+import { io, Socket } from "socket.io-client";
+import { getToken } from '@/utils/auth.utils';
 
-// export let socket: Socket | null = null;
+export const socketConnection = () => {
+  console.log("Start connection")
+  
+  const url = process.env.API_URL
+  if (!url) throw new Error("url is not defined");
+  
+  const roomId = "ea95d9c6-152b-4c85-ba1c-27a919edf13f";
+  
+  const socket: Socket = io(url, {
+    path: "/sqlizer/",
+    query: {
+      room: roomId,
+    },
+    transportOptions: {
+      polling: {
+        extraHeaders: {
+          BearerToken: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6ImNkMWJmZTUzLTk3ODItNDhjYy1iNjIyLTdlMTMyMjhiODRkNCIsImlhdCI6MTY5NzA2Njk1MywiZXhwIjoxNjk3NjcxNzUzfQ.gmls57_n5WLMablqeEr4IghF0gEwIZscfaxT16FA37k'
+        }
+      }
+    }
+  });
+  
+  console.log(socket)
+  socket?.on('socketError', (data: any) => console.log(data));  
+  console.log("Connection successful")
 
-// if(!socket) {
-//   console.log("Start connection")
-  
-//   const url = process.env.API_URL
-//   if (!url) throw new Error("url is not defined");
-  
-//   const roomId = "ea95d9c6-152b-4c85-ba1c-27a919edf13f";
-  
-//   socket = io(url, {
-//     path: "/sqlizer/",
-//     query: {
-//       room: roomId,
-//     },
-//     transportOptions: {
-//       polling: {
-//         extraHeaders: {
-//           BearerToken: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6ImNkMWJmZTUzLTk3ODItNDhjYy1iNjIyLTdlMTMyMjhiODRkNCIsImlhdCI6MTY5NjQ2MDQwNywiZXhwIjoxNjk3MDY1MjA3fQ.Mpqkni1e69t-dQ970R928xE1mqgPIFRtfOxxUIaJP3o'
-//         }
-//       }
-//     }
-//   });
-  
-//   console.log(socket)
-//   if (socket) {
-//     socket.on('socketError', (data) => console.log(data));
-
-//     socket.on('responseCreateTable', (data) => {
-//       console.log("----------------")
-//       console.log("SooS Salope")
-//       console.log(data)
-//     });
-  
-//     console.log("Connection successful")
-//     // console.log(socket)
-//   }
-// }
-
+  return socket
+}
