+++ src/sockets/socketEmitter.ts
@@ -3,5 +3,5 @@ import { socket } from '../components/Canvas/CanvasElement';
 
 export const addTableSocket = (table: Table) => {
   console.log(table)
-  socket?.emit("requestCreateTable", table);
+  socket.emit("requestCreateTable", table);
 };
